fix(proofs): extract shared Poseidon Merkle tree builder
Reduce SonarQube duplication by moving the full-tree Poseidon Merkle
builder into lib-proofs/src/merkle/poseidon_tree.rs and reusing it
from the state circuit.
