+++ lib-proofs/src/state/circuit.rs
@@ -239,43 +239,12 @@ pub mod real {
         leaves: &[[u64; NUM_HASH_OUT_ELTS]],
         leaf_index: usize,
     ) -> Result<([u64; NUM_HASH_OUT_ELTS], Vec<[u64; NUM_HASH_OUT_ELTS]>)> {
-        let max_leaves = 1usize << STATE_MERKLE_DEPTH;
-        if leaves.is_empty() || leaves.len() > max_leaves {
-            return Err(anyhow!(
-                "Leaf count {} out of range [1, {}]",
-                leaves.len(),
-                max_leaves
-            ));
-        }
-        if leaf_index >= leaves.len() {
-            return Err(anyhow!(
-                "leaf_index {} out of range [0, {})",
-                leaf_index,
-                leaves.len()
-            ));
-        }
-
-        let mut padded: Vec<Vec<F>> = leaves
-            .iter()
-            .map(|leaf| leaf.iter().map(|&v| F::from_canonical_u64(v)).collect())
-            .collect();
-        let target_len = padded.len().next_power_of_two();
-        while padded.len() < target_len {
-            padded.push(HashOut::from([F::ZERO; NUM_HASH_OUT_ELTS]).elements.to_vec());
-        }
-        while padded.len() < max_leaves {
-            padded.push(HashOut::from([F::ZERO; NUM_HASH_OUT_ELTS]).elements.to_vec());
-        }
-
-        let tree = MerkleTree::<F, PoseidonHash>::new(padded, 0);
-        let proof = tree.prove(leaf_index);
-        let root = tree.cap.0[0].elements.map(|f| f.to_canonical_u64());
-
-        let mut siblings = Vec::with_capacity(proof.siblings.len());
-        for s in &proof.siblings {
-            siblings.push(s.elements.map(|f| f.to_canonical_u64()));
-        }
-        Ok((root, siblings))
+        crate::merkle::poseidon_tree::build_poseidon_tree_from_hashes(
+            leaves,
+            leaf_index,
+            STATE_MERKLE_DEPTH,
+        )
+        .map_err(|e| anyhow!("State Merkle tree build failed: {}", e))
     }
 }
